pyproject.toml
[project]
name = "mcp-concept"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.8.0",
    "pydantic>=2.11.4",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "python-dotenv>=1.1.0",
]


[tool.pytest.ini_options]
python_files = "test_*.py"
asyncio_mode = "strict"


[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
